Get metadata and links to available digital elevation models

Get metadata and links to available digital elevation models

Usage

DEM_request(x)

Arguments

x

an sf, sfc or SpatVector object with one or more features (requests are based on the bounding boxes of the provided features)

Value

a data frame with metadata and links to the digital elevation models (different formats of digital terrain model, digital surface model and point clouds)

Details

The server can return a maximum of 2000 records in a single query. If your area of interest exceeds this limit, you can generate a grid of smaller polygons (sf::st_make_grid()) or a regular grid of points (sf::st_sample()).

See also

Examples

if (FALSE) { # \dontrun{
library(sf)
polygon_path = system.file("datasets/search_area.gpkg", package = "rgugik")
polygon = read_sf(polygon_path)
req_df = DEM_request(polygon)

# simple filtering by attributes
req_df = req_df[req_df$year > 2018, ]
req_df = req_df[req_df$product == "PointCloud" & req_df$format == "LAS", ]
} # }