Counties in Poland

The data frame contains the names of counties, their identifiers (TERYT, 4 characters) and the availability of building models in the LOD2 standard (logical value).

Usage

county_names

Format

An object of class data.frame with 380 rows and 3 columns.

Examples

county_names
#>                        NAME TERYT  LOD2
#> 1            aleksandrowski  0401  TRUE
#> 2               augustowski  2001  TRUE
#> 3               bartoszycki  2801  TRUE
#> 4              bełchatowski  1001 FALSE
#> 5                 będziński  2401  TRUE
#> 6                   bialski  0601  TRUE
#> 7            Biała Podlaska  0661  TRUE
#> 8              białobrzeski  1401  TRUE
#> 9              białogardzki  3201 FALSE
#> 10              białostocki  2002  TRUE
#> 11                Białystok  2061  TRUE
#> 12                  bielski  2003  TRUE
#> 13                  bielski  2402  TRUE
#> 14            Bielsko-Biała  2461  TRUE
#> 15      bieruńsko-lędziński  2414  TRUE
#> 16             bieszczadzki  1801  TRUE
#> 17              biłgorajski  0602  TRUE
#> 18                bocheński  1201  TRUE
#> 19            bolesławiecki  0201 FALSE
#> 20               braniewski  2802  TRUE
#> 21                brodnicki  0402  TRUE
#> 22                  brzeski  1202  TRUE
#> 23                  brzeski  1601  TRUE
#> 24               brzeziński  1021 FALSE
#> 25               brzozowski  1802  TRUE
#> 26                    buski  2601  TRUE
#> 27                 bydgoski  0403  TRUE
#> 28                Bydgoszcz  0461  TRUE
#> 29                    Bytom  2462  TRUE
#> 30                 bytowski  2201 FALSE
#> 31                    Chełm  0662  TRUE
#> 32               chełmiński  0404  TRUE
#> 33                 chełmski  0603  TRUE
#> 34               chodzieski  3001 FALSE
#> 35                chojnicki  2202 FALSE
#> 36                  Chorzów  2463  TRUE
#> 37             choszczeński  3202 FALSE
#> 38              chrzanowski  1203  TRUE
#> 39             ciechanowski  1402  TRUE
#> 40               cieszyński  2403  TRUE
#> 41  czarnkowsko-trzcianecki  3002 FALSE
#> 42              Częstochowa  2464  TRUE
#> 43            częstochowski  2404  TRUE
#> 44              człuchowski  2203 FALSE
#> 45         Dąbrowa Górnicza  2465  TRUE
#> 46                dąbrowski  1204  TRUE
#> 47                  dębicki  1803  TRUE
#> 48                  drawski  3203 FALSE
#> 49              działdowski  2803  TRUE
#> 50           dzierżoniowski  0202 FALSE
#> 51                   Elbląg  2861  TRUE
#> 52                 elbląski  2804  TRUE
#> 53                    ełcki  2805  TRUE
#> 54              garwoliński  1403  TRUE
#> 55                   Gdańsk  2261 FALSE
#> 56                  gdański  2204 FALSE
#> 57                   Gdynia  2262 FALSE
#> 58                  giżycki  2806  TRUE
#> 59                  Gliwice  2466  TRUE
#> 60                 gliwicki  2405  TRUE
#> 61                głogowski  0203 FALSE
#> 62               głubczycki  1602  TRUE
#> 63             gnieźnieński  3003 FALSE
#> 64              goleniowski  3204 FALSE
#> 65      golubsko-dobrzyński  0405  TRUE
#> 66                gołdapski  2818  TRUE
#> 67                 gorlicki  1205  TRUE
#> 68                gorzowski  0801 FALSE
#> 69      Gorzów Wielkopolski  0861 FALSE
#> 70              gostyniński  1404  TRUE
#> 71                gostyński  3004 FALSE
#> 72                 górowski  0204 FALSE
#> 73                grajewski  2004  TRUE
#> 74                grodziski  1405  TRUE
#> 75                grodziski  3005 FALSE
#> 76                 grójecki  1406  TRUE
#> 77                Grudziądz  0462  TRUE
#> 78              grudziądzki  0406  TRUE
#> 79                 gryficki  3205 FALSE
#> 80                gryfiński  3206 FALSE
#> 81                hajnowski  2005  TRUE
#> 82            hrubieszowski  0604  TRUE
#> 83                  iławski  2807  TRUE
#> 84            inowrocławski  0407  TRUE
#> 85                 janowski  0605  TRUE
#> 86               jarociński  3006 FALSE
#> 87              jarosławski  1804  TRUE
#> 88                jasielski  1805  TRUE
#> 89         Jastrzębie-Zdrój  2467  TRUE
#> 90                 jaworski  0205 FALSE
#> 91                 Jaworzno  2468  TRUE
#> 92             Jelenia Góra  0261 FALSE
#> 93            jeleniogórski  0206 FALSE
#> 94             jędrzejowski  2602  TRUE
#> 95                  kaliski  3007 FALSE
#> 96                   Kalisz  3061 FALSE
#> 97           kamiennogórski  0207 FALSE
#> 98                kamieński  3207 FALSE
#> 99                 kartuski  2205 FALSE
#> 100                Katowice  2469  TRUE
#> 101             kazimierski  2603  TRUE
#> 102 kędzierzyńsko-kozielski  1603  TRUE
#> 103                kępiński  3008 FALSE
#> 104              kętrzyński  2808  TRUE
#> 105                  Kielce  2661  TRUE
#> 106                kielecki  2604  TRUE
#> 107             kluczborski  1604  TRUE
#> 108                kłobucki  2406  TRUE
#> 109                 kłodzki  0208 FALSE
#> 110            kolbuszowski  1806  TRUE
#> 111               kolneński  2006  TRUE
#> 112                  kolski  3009 FALSE
#> 113             kołobrzeski  3208 FALSE
#> 114                 konecki  2605  TRUE
#> 115                   Konin  3062 FALSE
#> 116                koniński  3010 FALSE
#> 117                Koszalin  3261 FALSE
#> 118             koszaliński  3209 FALSE
#> 119              kościański  3011 FALSE
#> 120              kościerski  2206 FALSE
#> 121              kozienicki  1407  TRUE
#> 122               krakowski  1206  TRUE
#> 123                  Kraków  1261  TRUE
#> 124             krapkowicki  1605  TRUE
#> 125           krasnostawski  0606  TRUE
#> 126               kraśnicki  0607  TRUE
#> 127                  Krosno  1861  TRUE
#> 128             krośnieński  0802 FALSE
#> 129             krośnieński  1807  TRUE
#> 130            krotoszyński  3012 FALSE
#> 131               kutnowski  1002 FALSE
#> 132              kwidzyński  2207 FALSE
#> 133             legionowski  1408  TRUE
#> 134                 Legnica  0262 FALSE
#> 135                legnicki  0209 FALSE
#> 136                   leski  1821  TRUE
#> 137             leszczyński  3013 FALSE
#> 138                  Leszno  3063 FALSE
#> 139                leżajski  1808  TRUE
#> 140                lęborski  2208 FALSE
#> 141              lidzbarski  2809  TRUE
#> 142              limanowski  1207  TRUE
#> 143               lipnowski  0408  TRUE
#> 144                  lipski  1409  TRUE
#> 145             lubaczowski  1809  TRUE
#> 146                lubański  0210 FALSE
#> 147             lubartowski  0608  TRUE
#> 148                lubelski  0609  TRUE
#> 149                lubiński  0211 FALSE
#> 150                  Lublin  0663  TRUE
#> 151             lubliniecki  2407  TRUE
#> 152                lwówecki  0212 FALSE
#> 153                łańcucki  1810  TRUE
#> 154                   łaski  1003 FALSE
#> 155                łęczycki  1004 FALSE
#> 156               łęczyński  0610  TRUE
#> 157                 łobeski  3218 FALSE
#> 158                   Łomża  2062  TRUE
#> 159               łomżyński  2007  TRUE
#> 160                 łosicki  1410  TRUE
#> 161                 łowicki  1005 FALSE
#> 162         łódzki wschodni  1006 FALSE
#> 163                    Łódź  1061 FALSE
#> 164                łukowski  0611  TRUE
#> 165                makowski  1411  TRUE
#> 166               malborski  2209 FALSE
#> 167              miechowski  1208  TRUE
#> 168                mielecki  1811  TRUE
#> 169           międzychodzki  3014 FALSE
#> 170            międzyrzecki  0803 FALSE
#> 171              mikołowski  2408  TRUE
#> 172                 milicki  0213 FALSE
#> 173                  miński  1412  TRUE
#> 174                 mławski  1413  TRUE
#> 175              mogileński  0409  TRUE
#> 176                moniecki  2008  TRUE
#> 177               mrągowski  2810  TRUE
#> 178               Mysłowice  2470  TRUE
#> 179              myszkowski  2409  TRUE
#> 180              myślenicki  1209  TRUE
#> 181             myśliborski  3210 FALSE
#> 182               nakielski  0410  TRUE
#> 183             namysłowski  1606  TRUE
#> 184                nidzicki  2811  TRUE
#> 185                niżański  1812  TRUE
#> 186             nowodworski  1414  TRUE
#> 187             nowodworski  2210 FALSE
#> 188             nowomiejski  2812  TRUE
#> 189             nowosądecki  1210  TRUE
#> 190              nowosolski  0804 FALSE
#> 191              nowotarski  1211  TRUE
#> 192             nowotomyski  3015 FALSE
#> 193               Nowy Sącz  1262  TRUE
#> 194                   nyski  1607  TRUE
#> 195               obornicki  3016 FALSE
#> 196                  olecki  2813  TRUE
#> 197                  oleski  1608  TRUE
#> 198               oleśnicki  0214 FALSE
#> 199                 olkuski  1212  TRUE
#> 200                 Olsztyn  2862  TRUE
#> 201              olsztyński  2814  TRUE
#> 202                 oławski  0215 FALSE
#> 203               opatowski  2606  TRUE
#> 204              opoczyński  1007 FALSE
#> 205                   Opole  1661  TRUE
#> 206                 opolski  0612  TRUE
#> 207                 opolski  1609  TRUE
#> 208              ostrołęcki  1415  TRUE
#> 209               Ostrołęka  1461  TRUE
#> 210             ostrowiecki  2607  TRUE
#> 211               ostrowski  1416  TRUE
#> 212               ostrowski  3017 FALSE
#> 213               ostródzki  2815  TRUE
#> 214           ostrzeszowski  3018 FALSE
#> 215             oświęcimski  1213  TRUE
#> 216                 otwocki  1417  TRUE
#> 217              pabianicki  1008 FALSE
#> 218             pajęczański  1009 FALSE
#> 219              parczewski  0613  TRUE
#> 220            piaseczyński  1418  TRUE
#> 221         Piekary Śląskie  2471  TRUE
#> 222                  pilski  3019 FALSE
#> 223              pińczowski  2608  TRUE
#> 224             piotrkowski  1010 FALSE
#> 225    Piotrków Trybunalski  1062 FALSE
#> 226                   piski  2816  TRUE
#> 227              pleszewski  3020 FALSE
#> 228                   Płock  1462  TRUE
#> 229                  płocki  1419  TRUE
#> 230                 płoński  1420  TRUE
#> 231              poddębicki  1011 FALSE
#> 232                 policki  3211 FALSE
#> 233              polkowicki  0216 FALSE
#> 234                  Poznań  3064 FALSE
#> 235               poznański  3021 FALSE
#> 236             proszowicki  1214  TRUE
#> 237               prudnicki  1610  TRUE
#> 238             pruszkowski  1421  TRUE
#> 239              przasnyski  1422  TRUE
#> 240               przemyski  1813  TRUE
#> 241                Przemyśl  1862  TRUE
#> 242              przeworski  1814  TRUE
#> 243               przysuski  1423  TRUE
#> 244              pszczyński  2410  TRUE
#> 245                   pucki  2211 FALSE
#> 246                puławski  0614  TRUE
#> 247                pułtuski  1424  TRUE
#> 248                pyrzycki  3212 FALSE
#> 249              raciborski  2411  TRUE
#> 250                   Radom  1463  TRUE
#> 251                radomski  1425  TRUE
#> 252          radomszczański  1012 FALSE
#> 253            radziejowski  0411  TRUE
#> 254               radzyński  0615  TRUE
#> 255                 rawicki  3022 FALSE
#> 256                  rawski  1013 FALSE
#> 257  ropczycko-sędziszowski  1815  TRUE
#> 258             Ruda Śląska  2472  TRUE
#> 259                rybnicki  2412  TRUE
#> 260                  Rybnik  2473  TRUE
#> 261                   rycki  0616  TRUE
#> 262                rypiński  0412  TRUE
#> 263              rzeszowski  1816  TRUE
#> 264                 Rzeszów  1863  TRUE
#> 265            sandomierski  2609  TRUE
#> 266                 sanocki  1817  TRUE
#> 267               sejneński  2009  TRUE
#> 268              sępoleński  0413  TRUE
#> 269                 Siedlce  1464  TRUE
#> 270               siedlecki  1426  TRUE
#> 271    Siemianowice Śląskie  2474  TRUE
#> 272             siemiatycki  2010  TRUE
#> 273               sieradzki  1014 FALSE
#> 274               sierpecki  1427  TRUE
#> 275               skarżyski  2610  TRUE
#> 276            Skierniewice  1063 FALSE
#> 277           skierniewicki  1015 FALSE
#> 278              sławieński  3213 FALSE
#> 279                słubicki  0805 FALSE
#> 280                słupecki  3023 FALSE
#> 281                  Słupsk  2263 FALSE
#> 282                 słupski  2212 FALSE
#> 283            sochaczewski  1428  TRUE
#> 284              sokołowski  1429  TRUE
#> 285                sokólski  2011  TRUE
#> 286                   Sopot  2264 FALSE
#> 287               Sosnowiec  2475  TRUE
#> 288           stalowowolski  1818  TRUE
#> 289           starachowicki  2611  TRUE
#> 290             stargardzki  3214 FALSE
#> 291            starogardzki  2213 FALSE
#> 292              staszowski  2612  TRUE
#> 293              strzelecki  1611  TRUE
#> 294  strzelecko-drezdenecki  0806 FALSE
#> 295             strzeliński  0217 FALSE
#> 296             strzyżowski  1819  TRUE
#> 297              sulęciński  0807 FALSE
#> 298                   suski  1215  TRUE
#> 299                suwalski  2012  TRUE
#> 300                 Suwałki  2063  TRUE
#> 301             szamotulski  3024 FALSE
#> 302                Szczecin  3262 FALSE
#> 303            szczecinecki  3215 FALSE
#> 304            szczycieński  2817  TRUE
#> 305                sztumski  2216 FALSE
#> 306            szydłowiecki  1430  TRUE
#> 307                 średzki  0218 FALSE
#> 308                 średzki  3025 FALSE
#> 309                 śremski  3026 FALSE
#> 310               świdnicki  0219 FALSE
#> 311               świdnicki  0617  TRUE
#> 312              świdwiński  3216 FALSE
#> 313           świebodziński  0808 FALSE
#> 314                 świecki  0414  TRUE
#> 315          Świętochłowice  2476  TRUE
#> 316             Świnoujście  3263 FALSE
#> 317              Tarnobrzeg  1864  TRUE
#> 318            tarnobrzeski  1820  TRUE
#> 319             tarnogórski  2413  TRUE
#> 320               tarnowski  1216  TRUE
#> 321                  Tarnów  1263  TRUE
#> 322              tatrzański  1217  TRUE
#> 323                tczewski  2214 FALSE
#> 324             tomaszowski  0618  TRUE
#> 325             tomaszowski  1016 FALSE
#> 326                   Toruń  0463  TRUE
#> 327                toruński  0415  TRUE
#> 328              trzebnicki  0220 FALSE
#> 329               tucholski  0416  TRUE
#> 330                 turecki  3027 FALSE
#> 331                   Tychy  2477  TRUE
#> 332               wadowicki  1218  TRUE
#> 333               Wałbrzych  0265 FALSE
#> 334              wałbrzyski  0221 FALSE
#> 335                 wałecki  3217 FALSE
#> 336                Warszawa  1465  TRUE
#> 337     warszawski zachodni  1432  TRUE
#> 338               wąbrzeski  0417  TRUE
#> 339             wągrowiecki  3028 FALSE
#> 340             wejherowski  2215 FALSE
#> 341             węgorzewski  2819  TRUE
#> 342               węgrowski  1433  TRUE
#> 343                wielicki  1219  TRUE
#> 344               wieluński  1017 FALSE
#> 345            wieruszowski  1018 FALSE
#> 346               Włocławek  0464  TRUE
#> 347              włocławski  0418  TRUE
#> 348               włodawski  0619  TRUE
#> 349            włoszczowski  2613  TRUE
#> 350            wodzisławski  2415  TRUE
#> 351             wolsztyński  3029 FALSE
#> 352              wołomiński  1434  TRUE
#> 353                wołowski  0222 FALSE
#> 354                 Wrocław  0264 FALSE
#> 355              wrocławski  0223 FALSE
#> 356              wrzesiński  3030 FALSE
#> 357               wschowski  0812 FALSE
#> 358        wysokomazowiecki  2013  TRUE
#> 359              wyszkowski  1435  TRUE
#> 360                  Zabrze  2478  TRUE
#> 361              zambrowski  2014  TRUE
#> 362                zamojski  0620  TRUE
#> 363                  Zamość  0664  TRUE
#> 364           zawierciański  2416  TRUE
#> 365              ząbkowicki  0224 FALSE
#> 366           zduńskowolski  1019 FALSE
#> 367                zgierski  1020 FALSE
#> 368             zgorzelecki  0225 FALSE
#> 369            Zielona Góra  0862 FALSE
#> 370           zielonogórski  0809 FALSE
#> 371             złotoryjski  0226 FALSE
#> 372               złotowski  3031 FALSE
#> 373               zwoleński  1436  TRUE
#> 374                żagański  0810 FALSE
#> 375                  żarski  0811 FALSE
#> 376                 żniński  0419  TRUE
#> 377                    Żory  2479  TRUE
#> 378              żuromiński  1437  TRUE
#> 379             żyrardowski  1438  TRUE
#> 380                żywiecki  2417  TRUE