Download Land and Building Register (EGiB) layers

Download Land and Building Register (EGiB) layers

Usage

egib_download(
  county = NULL,
  TERYT = NULL,
  layer = "parcels",
  outdir = ".",
  ...
)

Arguments

county

County name in Polish. Check county_names() function.

TERYT

County ID (4 characters).

layer

Requested layer, parcels default. Other common layer is buildings. You can check available layers by egib_layers() dataset.

outdir

name of the output directory where the data has to be stored, current directory by default. If you don't want to download data, pass a NULL value.

...

any other parameter passed to sf::read_sf() function.

Value

If the outdir argument is specified, the data will be downloaded to disk in geopackage format. If the outdir argument is NULL, simple feature data frame is returned, but only for the first object.

References

The EGiB data (Ewidencja Gruntów i Budynków) consist of 2 primary layers: cadastral data (parcels) and buildings footprints. The data is maintained on county level, which results in 380 units (different sources of data). It may contain additional layers like points of detailed horizontal and vertical geodetic control network ("osnowa_pozioma" and "osnowa_pionowa" respectively).

https://www.geoportal.gov.pl/en/data/land-and-building-register-egib/

https://www.geoportal.gov.pl/en/data/detailed-control-network-database-bdsog/

Examples

if (FALSE) { # \dontrun{
egib_download(TERYT = c("2476", "2264"), layer = "buildings", outdir = ".") # 2.2 + 2.6 MB
parcels = egib_download(county = "Świętochłowice", layer = "parcels", outdir = NULL) # 3.9 MB
} # }