Download digital terrain models for voivodeships (100 m resolution)

Download digital terrain models for voivodeships (100 m resolution)

pointDTM100_download(voivodeships, outdir = ".", unzip = TRUE, ...)

Arguments

voivodeships

selected voivodeships in Polish or English, or TERC (function voivodeship_names() can by helpful)

outdir

(optional) name of the output directory; by default, files are saved in the working directory

unzip

TRUE (default) or FALSE, when TRUE the downloaded archive will be extracted and removed

...

additional argument for utils::download.file()

Value

text files with X, Y, Z columns (EPSG:2180)

Examples

if (FALSE) {
pointDTM100_download(c("opolskie", "świętokrzyskie")) # 8.5 MB
pointDTM100_download(c("Opole", "Swietokrzyskie")) # 8.5 MB
pointDTM100_download(c("16", "26")) # 8.5 MB
}