rgugik: Search and Retrieve Spatial Data from 'GUGiK'

Automatic open data acquisition from resources of Polish Head Office of Geodesy and Cartography ('Główny Urząd Geodezji i Kartografii') (https://www.gov.pl/web/gugik). Available datasets include various types of numeric, raster and vector data, such as orthophotomaps, digital elevation models (digital terrain models, digital surface model, point clouds), state register of borders, spatial databases, geometries of cadastral parcels, 3D models of buildings, and more. It is also possible to geocode addresses or objects using the geocodePL_get() function.

See also

Author

Maintainer: Krzysztof Dyba adres7@gmail.com (ORCID)

Authors:

Jakub Nowosad nowosad.jakub@gmail.com (ORCID)

Other contributors:

Maciej Beręsewicz maciej.beresewicz@ue.poznan.pl (ORCID) [contributor]

GUGiK (source of the data) [contributor]