Download Topographic Databases for counties

Download Topographic Databases for counties

Usage

topodb_download(county = NULL, TERYT = NULL, outdir = ".", unzip = TRUE, ...)

Arguments

county

county name in Polish. Check county_names() function.

TERYT

county ID (4 characters)

outdir

(optional) name of the output directory; by default, files are saved in the working directory

unzip

TRUE (default) or FALSE, when TRUE the downloaded archive will be extracted and removed

...

additional argument for utils::download.file()

Value

a database in Geography Markup Language format (.GML), the content and detail level corresponds to the topographic map in the scale of 1:10000

References

description of topographical and general geographical databases, and technical standards for making maps (in Polish): https://isap.sejm.gov.pl/isap.nsf/download.xsp/WDU20210001412/O/D20211412.pdf

brief description of categories and layer names (in English and Polish): https://kadyb.github.io/rgugik/articles/articles/spatialdb_description.html

Examples

if (FALSE) { # \dontrun{
topodb_download(county = "Świętochłowice") # 2.4 MB
topodb_download(TERYT = c("2476", "2264")) # 4.8 MB
} # }