Download General Geographic Databases for entire voivodeships

Download General Geographic Databases for entire voivodeships

geodb_download(voivodeships, outdir = ".", unzip = TRUE, ...)

Arguments

voivodeships

selected voivodeships in Polish or English, or TERC (object voivodeship_names can by helpful)

outdir

(optional) name of the output directory; by default, files are saved in the working directory

unzip

TRUE (default) or FALSE, when TRUE the downloaded archive will be extracted and removed

...

additional argument for utils::download.file()

Value

a database in Geography Markup Language format (.GML), the content and detail level corresponds to the general geographic map in the scale of 1:250000

References

description of topographical and general geographical databases, and technical standards for making maps (in Polish): https://isap.sejm.gov.pl/isap.nsf/download.xsp/WDU20210001412/O/D20211412.pdf

brief description of categories and layer names (in English and Polish): https://kadyb.github.io/rgugik/articles/articles/spatialdb_description.html

Examples

if (FALSE) { # \dontrun{
geodb_download(c("opolskie", "lubuskie")) # 12.7 MB
geodb_download(c("Opole", "Lubusz")) # 12.7 MB
geodb_download(c("16", "08")) # 12.7 MB
} # }